requirements.txt
streamlit
openai
pyyaml
sentence-transformers
